+++ src/components/general/CoinWrapperSmall/index.tsx
@@ -2,12 +2,12 @@ import SidebarDropdown from '../SidebarDropdown'
 import * as S from './style'
 
 
-const CoinWrapper =({CoinName,Cointype})=>{
+const CoinWrapper =({CoinName,CoinType})=>{
 
   return (<S.coinwrapper>
-            <S.CoinImage src={`img/cryptocurrencies/${Cointype}.png`} />
+            <S.CoinImage src={`img/cryptocurrencies/${CoinType}.png`} />
             <S.nameandvaluecolumn>
-                <span>{Cointype}</span>
+                <span>{CoinType}</span>
                 <S.blurtext>{CoinName}</S.blurtext>
             </S.nameandvaluecolumn>
           </S.coinwrapper>
