+++ src/components/general/SearchBar/index.tsx
@@ -5,11 +5,12 @@ export type SearchInputProps = {
     type: string
     value: string
     setValue: any
+    resize?:any
 }
 
-const SearchBar = ({placeholder, type, value, setValue}: SearchInputProps) => (
+const SearchBar = ({placeholder, type, value, setValue, resize = "NotActive"}: SearchInputProps) => (
     <S.Wrapper>
-        <S.Image src="/img/icons/Search.svg" />
+        <S.Image src="/img/icons/Search.svg"  className={resize}/>
         <S.Input type={type} placeholder={placeholder} value={value} onChange={(e) => setValue(e.target.value)} />
     </S.Wrapper>
 )
