+++ src/components/general/SidebarDropdown/index.tsx
@@ -1,19 +1,19 @@
 import * as S from './styles'
 
 export type SidebarDropdownProps = {
-  backgroundcolor: string;
+  backgroundColor: string;
   rotation:string;
-  hoverrotation:string;
+  hoverRotation:string;
   icon?: 'None' | 'ArrowTop' | 'ArrowDownBlack'
 }
 
 
 
-const SidebarDropdown = ({ rotation = 'rotate(270deg)', backgroundcolor='#1C2023', hoverrotation='rotate(0deg)' , icon='ArrowTop'}: SidebarDropdownProps) => {
+const SidebarDropdown = ({ rotation = 'rotate(270deg)', backgroundColor='#1C2023', hoverRotation='rotate(0deg)' , icon='ArrowTop'}: SidebarDropdownProps) => {
   return (
-    <S.Wrapper hoverrotation={hoverrotation}>
+    <S.Wrapper hoverRotation={hoverRotation}>
      
-        <S.WrapperImage backgroundcolor = {backgroundcolor}>
+        <S.WrapperImage backgroundColor = {backgroundColor}>
           <S.Image src={`/img/icons/${icon}.svg`} rotation = {rotation} />
         </S.WrapperImage>
