+++ src/templates/swap/style.ts
@@ -4,13 +4,11 @@ import { generateMedia } from 'styled-media-query'
 const customMedia = generateMedia({
     custom: '1199px'
   })
-  type Props = {
-    marketActive: boolean
-  }
+  
 
 export const Wrapper = styled.main`
   display: flex;
-  flex-direction:row;
+  flex-direction:column;
   height: 100vh;
   &.Active{
     height: 131vh;
@@ -33,6 +31,7 @@ margin-left:1.2rem;
   height: 8%;
 }
 `
+
 export const PolkaLogo = styled.img`
   width: 14%;
   margin-bottom: 2rem;
@@ -41,155 +40,15 @@ export const PolkaLogo = styled.img`
     height: 76%;
   }
 `
-export const ContentWrapper = styled.div`
-display:flex;
-flex-direction:column;
-width: 100%;
-height:100%;
-`
-export const BottomContentWrapper = styled.div`
-display:flex;
-flex-direction:row;
-width:100%;
-height:100%;
-background-color: #1C2023;
-`
-export const LeftColumn = styled.div`
-  display:flex;
-  flex-direction:column;
-  height:100%;
-  justify-content:flex-end;
-  margin-left:1.2rem;
-  padding-bottom:1.3rem;
-  flex:1
-`
-export const LabelWithIcon = styled.div`
-    display:flex;    
-    width: 224px;
-    height: 48px;
-    background: #2E303C 0% 0% no-repeat padding-box;
-    border-radius: 7px;
-    align-items:center;
-    justify-content:flex-start;
-    padding-left:0.8rem;
-    font-size:17px;
-    font-family: Work Sans,-apple-system,BlinkMacSystemFont,'Segoe UI',Roboto,Oxygen,Ubuntu,Cantarell,'Open Sans','Helvetica Neue',sans-serif;
-    font-weight:medium;
-    line-spacing:20px;
-`
-export const IconWrapper = styled.div`
-    padding-right:1rem;    
-`
-export const Iconcolumn = styled.div`
-    display:flex;
-    flex-direction:column;
-    justify-content:space-between;
-    margin-bottom:1rem;
-    
-`
-export const Marginbottom= styled.div`
-    margin-bottom:1rem;
-`
-export const MiddleColumn = styled.div`
-    display:flex;
-    flex-direction:column;
-    height:100%;
-    justify-content:flex-end;
-    margin-left:7rem;
-    flex:3;
-  
-`
-export const bottomlabel = styled.div`
-    
-    height: 4rem;
 
-    &.Active{
-      height:11rem;
-    }
-    display:flex;   
-    background: #24252C 0% 0% no-repeat padding-box;
-    box-shadow: 0px 100px 99px #0000002C;
-    border-radius: 12px 12px 0px 0px;
-    justify-content:center;
-    align-items:center;
-    font-size:17px;
-    font-family: Work Sans,-apple-system,BlinkMacSystemFont,'Segoe UI',Roboto,Oxygen,Ubuntu,Cantarell,'Open Sans','Helvetica Neue',sans-serif;
-    font-weight:400;
-    line-spacing:20px;
-`
-export const Bottomtextwrapper = styled.div`
-    display:flex;
-    flex-direction:row;
-    justify-content:space-evenly; 
-    width:17rem;
-`
-export const Image = styled.img`
-    width:10%;
-`
-export const RightColumn = styled.div`
-  display:flex;
-  flex-direction:column;
-  height:100%;
-  justify-content:flex-end;
-  margin-left:1.2rem;
-  padding-bottom:2rem;
-  flex:1;
-`
-export const bottomrightcontentwrapper = styled.div`
-display:flex;
-flex-direction:row;
-padding-left:15rem;
-align-items:center;
-`
-export const bottomcolor = styled.div`
-width: 10px;
-height: 10px;
-margin-right:1.5rem;
-background: #0CA564 0% 0% no-repeat padding-box;
-border-radius: 20px;
-`
+
 export const Root = styled.div`
   position: relative;
   background-color: #1C2023;
 
 `;
-export const ContainerActions = styled.div`
-  display:grid;
-  grid-template-columns: repeat(3, 1fr);
-  grid-column-gap:0.5rem;
-  margin-right:1rem;
-`
-export const ContainerTitle = styled.div`
-  display:flex;
-  flex-direction:row;
-  align-items:center;
-`
-export const Image2 = styled.img`
-  width:80%;
-  height:40%;
-  
-`
-export const ImageWrapper = styled.div`
-    background: #030517 0% 0% no-repeat ;
-    border-radius: 7px;
-    opacity: 1;
-    display: flex;
-    align-items:center;
-    justify-content: center;
-    cursor: pointer;    
-    padding-left:0.6rem;
-    padding-right:0.6rem;
-    
-    :hover {
-      opacity: 0.8;
-    }
-  
-`
-export const WrapperGraph = styled.div<Props>`
-  display: flex;
-  width:108%;
 
-`;
+
 export const Overlay = styled.div`
   position: fixed;
   top: 0;
@@ -198,18 +57,6 @@ export const Overlay = styled.div`
   right: 0;
   background: rgba(0, 0, 0, 0.3); 
 `;
-export const Dialog = styled.div`
-  border-radius: 5px;  
-  position: absolute;
-  height:80%;
-  top: 40%;
-  
-  transform: translate(-50%, -50%);
-  z-index: 1;
-  width:25%;
- 
-
-    `;
